requirements.txt
numpy
opencv-python
Pillow
onnxruntime
tqdm
scikit-image
fire
